refactor: split oversized test and fixture files
Extract ArmeriaFixtureTestBase PSI stubs into focused extension files
and split large route-analysis test classes by concern (duplicate index,
extended registration, virtual host, route decorator).
