--- plugin-route-analysis/src/test/kotlin/com/linecorp/intellij/plugins/armeria/explorer/ArmeriaKotlinExtendedRegistrationCollectorTest.kt
@@ -1,681 +1,0 @@
-package com.linecorp.intellij.plugins.armeria.explorer
-
-import com.linecorp.intellij.plugins.armeria.test.ArmeriaFixtureTestBase
-
-class ArmeriaKotlinExtendedRegistrationCollectorTest : ArmeriaFixtureTestBase() {
-
-    override fun registerArmeriaStubs() {
-        registerKotlinExtendedRegistrationStubs()
-    }
-
-    fun testCollectKotlinFileServiceRegistration() {
-        myFixture.configureByText(
-            "Main.kt",
-            """
-            package example
-
-            import com.linecorp.armeria.server.Server
-            import java.io.File
-
-            fun main() {
-                Server.builder()
-                    .fileService("/files/", File("/tmp"))
-                    .build()
-            }
-            """.trimIndent(),
-        )
-
-        val routes = ArmeriaRouteCollector.collect(project)
-        val fileRoute = routes.firstOrNull { it.routeMatch == RouteMatch.FILE_SERVICE }
-        assertNotNull(fileRoute)
-        assertEquals("/files/", fileRoute!!.path)
-    }
-
-    fun testCollectKotlinFluentRouteRegistration() {
-        myFixture.configureByText(
-            "Main.kt",
-            """
-            package example
-
-            import com.linecorp.armeria.server.Server
-
-            fun main() {
-                Server.builder()
-                    .route()
-                    .post("/api/items")
-                    .build(Any())
-                    .build()
-            }
-            """.trimIndent(),
-        )
-
-        val routes = ArmeriaRouteCollector.collect(project)
-        val fluentRoute = routes.firstOrNull { it.routeMatch == RouteMatch.ROUTE_FLUENT }
-        assertNotNull(fluentRoute)
-        assertEquals("POST", fluentRoute!!.httpMethod)
-        assertEquals("/api/items", fluentRoute.path)
-    }
-
-    fun testCollectKotlinChainedVirtualHostDoesNotAnnotatePrecedingServiceRoute() {
-        myFixture.configureByText(
-            "Main.kt",
-            """
-            package example
-
-            import com.linecorp.armeria.server.Server
-
-            fun main() {
-                Server.builder()
-                    .service("/api", ApiService())
-                    .virtualHost("api.example.com")
-                    .build()
-            }
-            """.trimIndent(),
-        )
-        myFixture.addClass("package example; public class ApiService {}")
-
-        val routes = ArmeriaRouteCollector.collect(project)
-        val serviceRoute = routes.firstOrNull { it.routeMatch == RouteMatch.SERVICE }
-        assertNotNull(serviceRoute)
-        assertEquals("/api", serviceRoute!!.path)
-        assertEquals("", serviceRoute.virtualHostName)
-    }
-
-    fun testCollectKotlinVirtualHostThenServiceAnnotatesServiceRoute() {
-        myFixture.configureByText(
-            "Main.kt",
-            """
-            package example
-
-            import com.linecorp.armeria.server.Server
-
-            fun main() {
-                Server.builder()
-                    .virtualHost("api.example.com")
-                    .service("/api", ApiService())
-                    .build()
-            }
-            """.trimIndent(),
-        )
-        myFixture.addClass("package example; public class ApiService {}")
-
-        val routes = ArmeriaRouteCollector.collect(project)
-        val serviceRoute = routes.firstOrNull { it.routeMatch == RouteMatch.SERVICE }
-        assertNotNull(serviceRoute)
-        assertEquals("/api", serviceRoute!!.path)
-        assertEquals("api.example.com", serviceRoute.virtualHostName)
-    }
-
-    fun testCollectKotlinNestedVirtualHostLambdaUsesInnerHostname() {
-        myFixture.configureByText(
-            "Main.kt",
-            """
-            package example
-
-            import com.linecorp.armeria.server.Server
-
-            fun main() {
-                Server.builder()
-                    .virtualHost("outer.example.com") { outer ->
-                        outer.virtualHost("inner.example.com") { inner ->
-                            inner.service("/api", ApiService())
-                        }
-                    }
-                    .build()
-            }
-            """.trimIndent(),
-        )
-        myFixture.addClass("package example; public class ApiService {}")
-
-        val routes = ArmeriaRouteCollector.collect(project)
-        val serviceRoute = routes.firstOrNull { it.routeMatch == RouteMatch.SERVICE }
-        assertNotNull(serviceRoute)
-        assertEquals("/api", serviceRoute!!.path)
-        assertEquals("inner.example.com", serviceRoute.virtualHostName)
-    }
-
-    fun testCollectKotlinDecoratorUnderRegistration() {
-        myFixture.configureByText(
-            "Main.kt",
-            """
-            package example
-
-            import com.linecorp.armeria.server.Server
-            import com.linecorp.armeria.server.logging.LoggingService
-
-            fun main() {
-                Server.builder()
-                    .decoratorUnder("/public", LoggingService.newDecorator())
-                    .build()
-            }
-            """.trimIndent(),
-        )
-
-        val routes = ArmeriaRouteCollector.collect(project)
-        val decoratorRoute = routes.firstOrNull { it.routeMatch == RouteMatch.DECORATOR_UNDER }
-        assertNotNull(decoratorRoute)
-        assertEquals("/public", decoratorRoute!!.path)
-    }
-
-    fun testCollectKotlinWithRouteRegistration() {
-        myFixture.configureByText(
-            "Main.kt",
-            """
-            package example
-
-            import com.linecorp.armeria.server.RouteBuilder
-            import com.linecorp.armeria.server.Server
-
-            fun main() {
-                Server.builder()
-                    .withRoute { route: RouteBuilder ->
-                        route.post("/wrapped").build(Any())
-                    }
-                    .build()
-            }
-            """.trimIndent(),
-        )
-
-        val routes = ArmeriaRouteCollector.collect(project)
-        val fluentRoutes = routes.filter { it.routeMatch == RouteMatch.ROUTE_FLUENT }
-        assertEquals(1, fluentRoutes.size)
-        assertEquals("POST", fluentRoutes.single().httpMethod)
-        assertEquals("/wrapped", fluentRoutes.single().path)
-    }
-
-    fun testCollectKotlinWithRouteDoesNotBurnDedupKeyOnInvalidLambda() {
-        myFixture.configureByText(
-            "Main.kt",
-            """
-            package example
-
-            import com.linecorp.armeria.server.RouteBuilder
-            import com.linecorp.armeria.server.Server
-
-            fun main() {
-                Server.builder()
-                    .withRoute { _: RouteBuilder -> null }
-                    .withRoute { route: RouteBuilder ->
-                        route.post("/wrapped").build(Any())
-                    }
-                    .build()
-            }
-            """.trimIndent(),
-        )
-
-        val routes = ArmeriaRouteCollector.collect(project)
-        val fluentRoutes = routes.filter { it.routeMatch == RouteMatch.ROUTE_FLUENT }
-        assertEquals(1, fluentRoutes.size)
-        assertEquals("/wrapped", fluentRoutes.single().path)
-    }
-
-    fun testCollectKotlinHealthCheckRegistration() {
-        myFixture.configureByText(
-            "Main.kt",
-            """
-            package example
-
-            import com.linecorp.armeria.server.Server
-
-            fun main() {
-                Server.builder()
-                    .healthCheckService()
-                    .build()
-            }
-            """.trimIndent(),
-        )
-
-        val routes = ArmeriaRouteCollector.collect(project)
-        val healthRoute = routes.firstOrNull { it.routeMatch == RouteMatch.HEALTH_CHECK }
-        assertNotNull(healthRoute)
-        assertEquals("/internal/healthcheck", healthRoute!!.path)
-    }
-
-    fun testCollectKotlinVirtualHostRegistration() {
-        myFixture.configureByText(
-            "Main.kt",
-            """
-            package example
-
-            import com.linecorp.armeria.server.Server
-
-            fun main() {
-                Server.builder()
-                    .virtualHost("api.example.com")
-                    .build()
-            }
-            """.trimIndent(),
-        )
-
-        val routes = ArmeriaRouteCollector.collect(project)
-        val virtualHostRoute = routes.firstOrNull { it.routeMatch == RouteMatch.VIRTUAL_HOST }
-        assertNotNull(virtualHostRoute)
-        assertEquals("api.example.com", virtualHostRoute!!.virtualHostName)
-    }
-
-    fun testCollectKotlinRouteDecoratorDefaultPathTypeIsGlob() {
-        myFixture.configureByText(
-            "Main.kt",
-            """
-            package example
-
-            import com.linecorp.armeria.server.Server
-            import com.linecorp.armeria.server.logging.LoggingService
-
-            fun main() {
-                Server.builder()
-                    .routeDecorator()
-                    .build(LoggingService.newDecorator())
-                    .build()
-            }
-            """.trimIndent(),
-        )
-
-        val routes = ArmeriaRouteCollector.collect(project)
-        val decoratorRoute = routes.firstOrNull { it.routeMatch == RouteMatch.ROUTE_DECORATOR }
-        assertNotNull(decoratorRoute)
-        assertEquals(PathType.GLOB, decoratorRoute!!.pathType)
-        assertEquals("/**", decoratorRoute.path)
-    }
-
-    fun testCollectKotlinRouteDecoratorRegistration() {
-        myFixture.configureByText(
-            "Main.kt",
-            """
-            package example
-
-            import com.linecorp.armeria.server.Server
-            import com.linecorp.armeria.server.logging.LoggingService
-
-            fun main() {
-                Server.builder()
-                    .routeDecorator()
-                    .path("/decorated/**")
-                    .build(LoggingService.newDecorator())
-                    .build()
-            }
-            """.trimIndent(),
-        )
-
-        val routes = ArmeriaRouteCollector.collect(project)
-        val decoratorRoute = routes.firstOrNull { it.routeMatch == RouteMatch.ROUTE_DECORATOR }
-        assertNotNull(decoratorRoute)
-        assertEquals(PathType.EXACT, decoratorRoute!!.pathType)
-        assertEquals("/decorated/**", decoratorRoute.path)
-    }
-
-    fun testCollectKotlinRouteDecoratorMethodsStripHttpMethodPrefix() {
-        myFixture.configureByText(
-            "Main.kt",
-            """
-            package example
-
-            import com.linecorp.armeria.common.HttpMethod
-            import com.linecorp.armeria.server.Server
-            import com.linecorp.armeria.server.logging.LoggingService
-
-            fun main() {
-                Server.builder()
-                    .routeDecorator()
-                    .methods(HttpMethod.POST, HttpMethod.PUT)
-                    .build(LoggingService.newDecorator())
-                    .build()
-            }
-            """.trimIndent(),
-        )
-
-        val routes = ArmeriaRouteCollector.collect(project)
-        val decoratorRoute = routes.firstOrNull { it.routeMatch == RouteMatch.ROUTE_DECORATOR }
-        assertNotNull(decoratorRoute)
-        assertEquals("POST, PUT", decoratorRoute!!.httpMethod)
-    }
-
-    fun testCollectKotlinRouteDecoratorIgnoresPathCallsInsideDecoratorLambda() {
-        myFixture.configureByText(
-            "Main.kt",
-            """
-            package example
-
-            import com.linecorp.armeria.server.Server
-            import com.linecorp.armeria.server.logging.LoggingService
-
-            fun main() {
-                Server.builder()
-                    .routeDecorator()
-                    .path("/decorated/**")
-                    .build(decoratorWithNestedPath(LoggingService.newDecorator()))
-                    .build()
-            }
-
-            private fun decoratorWithNestedPath(decorator: Any): Any {
-                LoggingService.builder().path("/wrong/**").build()
-                return decorator
-            }
-            """.trimIndent(),
-        )
-
-        val routes = ArmeriaRouteCollector.collect(project)
-        val decoratorRoute = routes.firstOrNull { it.routeMatch == RouteMatch.ROUTE_DECORATOR }
-        assertNotNull(decoratorRoute)
-        assertEquals("/decorated/**", decoratorRoute!!.path)
-    }
-
-    fun testIgnoreNonArmeriaKotlinFluentBuildCalls() {
-        myFixture.configureByText(
-            "Main.kt",
-            """
-            package example
-
-            fun main() {
-                OtherBuilder.builder()
-                    .route()
-                    .post("/nope")
-                    .build(Any())
-                    .build()
-            }
-            """.trimIndent(),
-        )
-
-        val routes = ArmeriaRouteCollector.collect(project)
-        assertTrue(routes.none { it.routeMatch == RouteMatch.ROUTE_FLUENT })
-    }
-
-    fun testIgnoreNonArmeriaServiceInVirtualHostLambda() {
-        myFixture.configureByText(
-            "Main.kt",
-            """
-            package example
-
-            import com.linecorp.armeria.server.Server
-
-            fun main() {
-                Server.builder()
-                    .virtualHost("api.example.com") { vh ->
-                        OtherBuilder.builder()
-                            .service("/nope", Any())
-                            .build()
-                    }
-                    .build()
-            }
-            """.trimIndent(),
-        )
-
-        val routes = ArmeriaRouteCollector.collect(project)
-        assertTrue(routes.none { it.routeMatch == RouteMatch.SERVICE })
-    }
-
-    fun testCollectKotlinVirtualHostLambdaAnnotatesFileService() {
-        myFixture.configureByText(
-            "Main.kt",
-            """
-            package example
-
-            import com.linecorp.armeria.server.Server
-            import java.io.File
-
-            fun main() {
-                Server.builder()
-                    .virtualHost("api.example.com") { sb ->
-                        sb.fileService("/files/", File("/tmp"))
-                    }
-                    .build()
-            }
-            """.trimIndent(),
-        )
-
-        val routes = ArmeriaRouteCollector.collect(project)
-        val fileRoute = routes.firstOrNull { it.routeMatch == RouteMatch.FILE_SERVICE }
-        assertNotNull(fileRoute)
-        assertEquals("/files/", fileRoute!!.path)
-        assertEquals("api.example.com", fileRoute.virtualHostName)
-    }
-
-    fun testCollectKotlinFluentRoutePathPrefix() {
-        myFixture.configureByText(
-            "Main.kt",
-            """
-            package example
-
-            import com.linecorp.armeria.server.Server
-
-            fun main() {
-                Server.builder()
-                    .route()
-                    .pathPrefix("/api")
-                    .get("/items")
-                    .build(Any())
-                    .build()
-            }
-            """.trimIndent(),
-        )
-
-        val routes = ArmeriaRouteCollector.collect(project)
-        val fluentRoute = routes.firstOrNull { it.routeMatch == RouteMatch.ROUTE_FLUENT }
-        assertNotNull(fluentRoute)
-        assertEquals(PathType.EXACT, fluentRoute!!.pathType)
-        assertEquals("/api/items", fluentRoute.path)
-        assertEquals("GET", fluentRoute.httpMethod)
-    }
-
-    fun testCollectKotlinFileServiceFromConstValPath() {
-        myFixture.configureByText(
-            "Main.kt",
-            """
-            package example
-
-            import com.linecorp.armeria.server.Server
-            import java.io.File
-
-            private const val FILES_PATH = "/files/"
-
-            fun main() {
-                Server.builder()
-                    .fileService(FILES_PATH, File("/tmp"))
-                    .build()
-            }
-            """.trimIndent(),
-        )
-
-        val routes = ArmeriaRouteCollector.collect(project)
-        val fileRoute = routes.firstOrNull { it.routeMatch == RouteMatch.FILE_SERVICE }
-        assertNotNull(fileRoute)
-        assertEquals("/files/", fileRoute!!.path)
-    }
-
-    fun testIgnoreNonArmeriaFluentBuildInVirtualHostLambda() {
-        myFixture.configureByText(
-            "Main.kt",
-            """
-            package example
-
-            import com.linecorp.armeria.server.Server
-
-            fun main() {
-                Server.builder()
-                    .virtualHost("api.example.com") {
-                        OtherBuilder.builder()
-                            .route()
-                            .post("/nope")
-                            .build(Any())
-                            .build()
-                    }
-                    .build()
-            }
-            """.trimIndent(),
-        )
-
-        val routes = ArmeriaRouteCollector.collect(project)
-        assertTrue(routes.none { it.routeMatch == RouteMatch.ROUTE_FLUENT })
-    }
-
-    private fun registerKotlinExtendedRegistrationStubs() {
-        myFixture.addClass(
-            """
-            package com.linecorp.armeria.server;
-
-            public final class Server {
-                public static ServerBuilder builder() {
-                    return null;
-                }
-            }
-            """.trimIndent(),
-        )
-        myFixture.addClass(
-            """
-            package com.linecorp.armeria.server;
-
-            public final class ServerBuilder {
-                public ServerBuilder service(String path, Object service) {
-                    return this;
-                }
-
-                public ServerBuilder fileService(String path, java.io.File root) {
-                    return this;
-                }
-
-                public ServerBuilder healthCheckService() {
-                    return this;
-                }
-
-                public ServerBuilder virtualHost(String hostname) {
-                    return this;
-                }
-
-                public ServerBuilder virtualHost(String hostname, java.util.function.Consumer<ServerBuilder> customizer) {
-                    return this;
-                }
-
-                public ServerBuilder routeDecorator() {
-                    return this;
-                }
-
-                public ServerBuilder path(String pathPattern) {
-                    return this;
-                }
-
-                public ServerBuilder methods(Object... methods) {
-                    return this;
-                }
-
-                public ServerBuilder withRoute(java.util.function.Function<RouteBuilder, RouteBuilder> fn) {
-                    return this;
-                }
-
-                public ServerBuilder route() {
-                    return this;
-                }
-
-                public ServerBuilder post(String path) {
-                    return this;
-                }
-
-                public ServerBuilder get(String path) {
-                    return this;
-                }
-
-                public ServerBuilder pathPrefix(String pathPrefix) {
-                    return this;
-                }
-
-                public ServerBuilder build(Object handler) {
-                    return this;
-                }
-
-                public ServerBuilder decoratorUnder(String path, Object decorator) {
-                    return this;
-                }
-
-                public com.linecorp.armeria.server.Server build() {
-                    return null;
-                }
-            }
-            """.trimIndent(),
-        )
-        myFixture.addClass(
-            """
-            package com.linecorp.armeria.server;
-
-            public final class RouteBuilder {
-                public RouteBuilder post(String path) {
-                    return this;
-                }
-
-                public RouteBuilder build(Object handler) {
-                    return this;
-                }
-            }
-            """.trimIndent(),
-        )
-        myFixture.addClass(
-            """
-            package com.linecorp.armeria.server.logging;
-
-            public final class LoggingService {
-                public static LoggingServiceBuilder builder() {
-                    return null;
-                }
-
-                public static Object newDecorator() {
-                    return null;
-                }
-            }
-            """.trimIndent(),
-        )
-        myFixture.addClass(
-            """
-            package com.linecorp.armeria.server.logging;
-
-            public final class LoggingServiceBuilder {
-                public LoggingServiceBuilder path(String pathPattern) {
-                    return this;
-                }
-
-                public Object build() {
-                    return null;
-                }
-            }
-            """.trimIndent(),
-        )
-        myFixture.addClass(
-            """
-            package com.linecorp.armeria.common;
-
-            public final class HttpMethod {
-                public static final HttpMethod POST = null;
-                public static final HttpMethod PUT = null;
-            }
-            """.trimIndent(),
-        )
-        myFixture.addClass(
-            """
-            package example;
-
-            public final class OtherBuilder {
-                public static OtherBuilder builder() {
-                    return null;
-                }
-
-                public OtherBuilder route() {
-                    return this;
-                }
-
-                public OtherBuilder post(String path) {
-                    return this;
-                }
-
-                public OtherBuilder build(Object handler) {
-                    return this;
-                }
-
-                public OtherBuilder service(String path, Object handler) {
-                    return this;
-                }
-
-                public Object build() {
-                    return null;
-                }
-            }
-            """.trimIndent(),
-        )
-    }
-}
